Remove higher tracking features from the game
The game will no longer track which cards were touched by which clues.
The player will have to keep track of that themselves if needed.
The CardQuantum can be used for that (as it is used by the hyphenate
player). Most decision can be done now without calling up to the game
again. This reduces to need for the game reference (that sould be
reduced so that the player references can again be added to the game
struct without causing borrowing issues).